refactor: move components to sharedComp and refactor PatientDetails
