+++ src/components/dashboard/styles/Dashboard.style.ts
@@ -1,5 +1,7 @@
 import { createStyles, Theme } from "@material-ui/core/styles";
 
+//TODO clean up this style sheet
+
 const styles = (theme: Theme) =>
   createStyles({
     root: {
