+++ src/components/patientsDatabase/styles/PatientDetails.style.ts
