+++ src/components/sharedComponents/styles/AppointmentsItem.style.ts
